+++ src/ECommerce.Business/Services/Interfaces/IInventoryLogService.cs
@@ -4,6 +4,6 @@ namespace ECommerce.Business.Services.Interfaces
 {
     public interface IInventoryLogService
     {
-        Task WriteAsync(int productId, string action, int quantity, int oldStock, int newStock, string? referenceId);
+        Task WriteAsync(int productId, string action, decimal quantity, decimal oldStock, decimal newStock, string? referenceId);
     }
 }
